api/requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==21.4.0
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
dnspython==2.2.1
fastapi==0.78.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyCLI==2.0.3
pycodestyle==2.8.0
pydantic==1.9.1
pymongo==4.1.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
requests==2.28.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
